3rd party package: inquirer
```
$ npm install inquirer@7.3.1
$ npm install --save-dev @types/inquirer@7.3.1
$ npm ls
todo@1.0.0 /home/feri/wrk/ts-self-study/todo
├── @types/inquirer@7.3.1
└── inquirer@7.3.1
```

Chose 7.3.1 because that seems to be the latest version available
in the [Definitely Typed](https://github.com/DefinitelyTyped/DefinitelyTyped)
project (@types/inquirer).

The quality of the error message and the fact that had to downgrade to several
months old version of inquirer makes it not look so nice, though.